requirements.txt
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tqdm==4.47.0
Pillow==7.2.0
numpy==1.18.5
matplotlib==3.2.2
pyyaml==5.3.1
tensorboardx==2.1
h5py==3.1.0
